Removed duplicate file
